requirements.txt
Babel==2.9.1
blinker==1.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
PyMySQL==1.0.2
python-dotenv==0.17.1
pytz==2021.1
speaklater==1.3
typing-extensions==3.10.0.0
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
